fixed hartnell console positions (for now, still have to implement trights new model and new positions :sob:) and made the screenshake a little less crazy, and made it do it on flight tick as well.

+++ src/main/java/mdteam/ait/client/util/ClientShakeUtil.java
@@ -3,7 +3,7 @@ package mdteam.ait.client.util;
 import net.minecraft.client.MinecraftClient;
 
 public class ClientShakeUtil {
-    public static void shake(double scale) {
+    public static void shake(float scale) {
         if (MinecraftClient.getInstance().player == null) return;
 
         MinecraftClient.getInstance().player.setPitch(getShakeX(scale));
@@ -11,28 +11,28 @@ public class ClientShakeUtil {
     }
 
     /**
-     * gets the current yaw and adds an amount of a random from 0 - 1 timesed by the scale
+     * gets the current yaw and adds an amount of a random from 0 to 1 multiplied by the scale
      * @param scale the intensity of the shake
      * @return the new yaw
      */
-    private static float getShakeY(double scale) {
+    private static float getShakeY(float scale) {
         if (MinecraftClient.getInstance().player == null) return 0;
 
-        var temp = (MinecraftClient.getInstance().player.getRandom().nextFloat() * (float) scale);
+        var temp = (MinecraftClient.getInstance().player.getRandom().nextFloat() * scale);
 
         // this is a long line but i have a vendetta against creating variable for some reason
         return MinecraftClient.getInstance().player.getYaw() + (MinecraftClient.getInstance().player.getRandom().nextBoolean() ? temp : -temp);
     }
 
     /**
-     * gets the current pitch and adds an amount of a random from 0 - 1 timesed by the scale
+     * gets the current pitch and adds an amount of a random from 0 to 1 multiplied by the scale
      * @param scale the intensity of the shake
      * @return the new pitch
      */
-    private static float getShakeX(double scale) {
+    private static float getShakeX(float scale) {
         if (MinecraftClient.getInstance().player == null) return 0;
 
-        var temp = (MinecraftClient.getInstance().player.getRandom().nextFloat() * (float) scale);
+        var temp = (MinecraftClient.getInstance().player.getRandom().nextFloat() * scale);
 
         // yeah
         return MinecraftClient.getInstance().player.getPitch() + (MinecraftClient.getInstance().player.getRandom().nextBoolean() ? temp : -temp);
